backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
celery>=5.3,<6.0
redis>=5.0,<6.0
psycopg2-binary>=2.9,<3.0
python-decouple>=3.8,<4.0
deepface>=0.0.93
tf-keras>=2.16
Pillow>=10.0,<11.0
boto3>=1.34,<2.0
cryptography>=42.0,<43.0

# Dev / testes
ruff>=0.4
pytest>=8.0
pytest-django>=4.8
